• Source: https://github.com/osm0sis/AnyKernel3
• Branch: master

• Commit log (13 commits):
  AK3+Backend: add preliminary KernelSU support
  AK3: don't attempt to AVBv1 sign if /system/bin/dalvikvm is not present
  LICENSE: update XDA link
  README: update for snapshotupdater_static
  Backend: fix loop mounts on newer Lineage-based recoveries
  tools: newer busybox (1.36.1) patched to fix loop mounts on kernels <5.x
  Backend: abort apex loop mounts early if failing
  Backend: fixup now that mount_apex direct .img mount works
  tools: update to latest busybox (1.36.1)
  Backend: move busybox bin setup to setup_bb function, add tools to PATH
  AK3: clear cow before logical partition replace or resize (#78)
  AK3+Backend: pass AKHOME via export
  Backend: simpify getprop replacement propval return

• Conflicts:
	META-INF/com/google/android/update-binary
	tools/ak3-core.sh

+++ tools/ak3-core.sh
@@ -4,6 +4,7 @@
 [ "$OUTFD" ] || OUTFD=$1;
 
 # set up working directory variables
+[ "$AKHOME" ] && home=$AKHOME;
 [ "$home" ] || home=$PWD;
 bootimg=$home/boot.img;
 bin=$home/tools;
@@ -346,6 +347,28 @@ flash_boot() {
           for fdt in dtb extra kernel_dtb recovery_dtbo; do
             [ -f $fdt ] && $bin/magiskboot dtb $fdt patch; # remove dtb verity/avb
           done;
+        elif [ -d /data/adb/ksu -a -f /data/adb/ksu/modules.img ] && [ "$(file_getprop $home/anykernel.sh do.systemless)" == 1 ]; then
+          ui_print " " "KernelSU detected! Setting up for kernel helper module...";
+          comp=$($bin/magiskboot decompress kernel 2>&1 | grep -vE 'raw|zimage' | sed -n 's;.*\[\(.*\)\];\1;p');
+          ($bin/magiskboot split $kernel || $bin/magiskboot decompress $kernel kernel) 2>/dev/null;
+          if [ $? != 0 -a "$comp" ] && $comp --help 2>/dev/null; then
+            echo "Attempting kernel unpack with busybox $comp..." >&2;
+            $comp -dc $kernel > kernel;
+          fi;
+          if strings kernel 2>/dev/null | grep -q -E '^/data/adb/ksud$'; then
+            touch $home/kernelsu_patched;
+            strings kernel 2>/dev/null | grep -E -m1 'Linux version.*#' > $home/vertmp;
+          else
+            ui_print " " "Warning: No KernelSU support detected in kernel!";
+          fi;
+          if [ "$comp" ]; then
+            $bin/magiskboot compress=$comp kernel kernel.$comp;
+            if [ $? != 0 ] && $comp --help 2>/dev/null; then
+              echo "Attempting kernel repack with busybox $comp..." >&2;
+              $comp -9c kernel > kernel.$comp;
+            fi;
+            mv -f kernel.$comp kernel;
+          fi;
         else
           case $kernel in
             *-dtb) rm -f kernel_dtb;;
@@ -385,7 +408,7 @@ flash_boot() {
       *recovery*|*SOS*) avbtype=recovery;;
       *) avbtype=boot;;
     esac;
-    if [ "$(/system/bin/dalvikvm -Xnoimage-dex2oat -cp $bin/boot_signer-dexed.jar com.android.verity.BootSignature -verify boot.img 2>&1 | grep VALID)" ]; then
+    if [ -f /system/bin/dalvikvm ] && [ "$(/system/bin/dalvikvm -Xnoimage-dex2oat -cp $bin/boot_signer-dexed.jar com.android.verity.BootSignature -verify boot.img 2>&1 | grep VALID)" ]; then
       echo "Signing with AVBv1..." >&2;
       /system/bin/dalvikvm -Xnoimage-dex2oat -cp $bin/boot_signer-dexed.jar com.android.verity.BootSignature /$avbtype boot-new.img $pk8 $cert boot-new-signed.img;
     fi;
@@ -465,6 +488,8 @@ flash_generic() {
       fi
       echo "Removing any existing $1_ak3..." >&2;
       $bin/lptools_static remove $1_ak3;
+      echo "Clearing any merged cow partitions..." >&2;
+      $bin/lptools_static clear-cow;
       echo "Attempting to create $1_ak3..." >&2;
       if $bin/lptools_static create $1_ak3 $(wc -c < $img); then
         echo "Replacing $1$slot with $1_ak3..." >&2;
